strategy-service/requirements.txt
# 基础依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 数据处理
numpy==1.26.3
pandas==2.1.4

# 技术指标
# TA-Lib 安装复杂，暂时注释
# TA-Lib==0.4.26

# 安全
restrictedpython==8.1

# 记忆系统
rank-bm25==0.2.2

# 日志
python-json-logger==2.0.7

# 测试
pytest==8.3.3
